\name{CGIwithR-internal}

\title{Functions and Data Used Internally by the CGIwithR Package}

\alias{CGIparse}
\alias{ascii}
\alias{hexDecode}

\description{
  Functions used in decoding the CGI query string
  }

\author{David Firth, \email{david.firth@nuffield.ox.ac.uk}}

\keyword{internal}

\examples{}
